Update Conceptual Design.md
Updated for this feedback:

Comparative Analysis of Potential Solutions is still lacking defined metrics with data and citations to reinforce the argument. You need to clearly define the metrics of importance for each solution and state directly using cited data or reasonable arguments how every solution compares. For example, if cost is the most important factor for a solution, you need to state the cost for every potential solution so it is clear why you chose the subsystem solution.
Please add the total overall price of all subsystems combined to the budget section. Also, are all these items going to be purchased or are there some items that will be inherited from the previous group? If so, I would note that in the budget section.
On your Gantt chart, please make sure to update the progress column to include either completed, in progress, or planned